+++ server/src/main/java/org/allaymc/server/block/component/BlockEntityHolderComponentImpl.java
@@ -5,13 +5,17 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.allaymc.api.block.component.BlockEntityHolderComponent;
 import org.allaymc.api.blockentity.BlockEntity;
+import org.allaymc.api.blockentity.initinfo.BlockEntityInitInfo;
 import org.allaymc.api.blockentity.type.BlockEntityType;
 import org.allaymc.api.eventbus.EventHandler;
 import org.allaymc.api.math.position.Position3i;
+import org.allaymc.api.math.position.Position3ic;
 import org.allaymc.api.utils.identifier.Identifier;
 import org.allaymc.server.block.component.event.*;
 import org.allaymc.server.blockentity.component.BlockEntityBaseComponentImpl;
 import org.allaymc.server.blockentity.impl.BlockEntityImpl;
+
+import java.util.Objects;
 
 /**
  * @author daoge_cmd
@@ -25,20 +29,55 @@ public class BlockEntityHolderComponentImpl<T extends BlockEntity> implements Bl
     @Getter
     protected final BlockEntityType<T> blockEntityType;
 
+    /**
+     * Create block entity at a specific location.
+     *
+     * @param pos block entity's pos
+     */
+    protected void createBlockEntity(Position3ic pos) {
+        var dimension = pos.dimension();
+        Objects.requireNonNull(dimension);
+        var chunk = dimension.getChunkManager().getChunkByDimensionPos(pos.x(), pos.z());
+        if (chunk == null) {
+            throw new IllegalStateException("Trying to create a block entity in an unload chunk! Dimension: " + dimension + " at pos " + pos);
+        }
+        var presentBlockEntity = chunk.getBlockEntity(pos.x() & 15, pos.y(), pos.z() & 15);
+        if (presentBlockEntity != null) {
+            throw new IllegalStateException("Trying to create a block entity when block entity already exists! Dimension: " + dimension + " at pos " + pos);
+        }
+        var blockEntity = getBlockEntityType().createBlockEntity(BlockEntityInitInfo.builder().pos(pos.x(), pos.y(), pos.z()).dimension(dimension).build());
+        chunk.addBlockEntity(blockEntity);
+    }
+
+    /**
+     * Remove a block entity in a specific location.
+     *
+     * @param pos block entity's pos
+     * @throws IllegalStateException if chunk isn't loaded or the block entity not exists
+     */
+    protected void removeBlockEntity(Position3ic pos) {
+        var dimension = pos.dimension();
+        Objects.requireNonNull(dimension);
+        var chunk = dimension.getChunkManager().getChunkByDimensionPos(pos.x(), pos.z());
+        if (chunk == null) {
+            throw new IllegalStateException("Trying to remove a block entity in an unload chunk! Dimension: " + dimension + " at pos " + pos);
+        }
+        if (chunk.removeBlockEntity(pos.x() & 15, pos.y(), pos.z() & 15) == null) {
+            throw new IllegalStateException("Trying to remove a block entity which is not exists in Dimension " + dimension + " at pos " + pos);
+        }
+    }
+
     @EventHandler
     protected void onBlockPlace(CBlockOnPlaceEvent event) {
         var pos = event.getCurrentBlock().getPosition();
-
-        createBlockEntity(pos, false);
+        createBlockEntity(pos);
         var blockEntity = getBlockEntity(pos);
-        getBaseComponentImpl(blockEntity).onBlockPlace(event);
-
+        var baseComponent = getBaseComponentImpl(blockEntity);
+        baseComponent.onBlockPlace(event);
         // Send block entity to client after called onPlace() because onPlace() method may make some changes
-        // on this block entity
-        if (blockEntity.sendToClient()) {
-            // Send the block entity in the next chunk tick because we need to wait for the block being sent first
-            blockEntity.sendBlockEntityToViewers(false);
-        }
+        // on this block entity. Also, we should send the block entity in the next chunk tick because we need
+        // to wait for the block being sent first
+        baseComponent.sendBlockEntityToViewers(false);
     }
 
     @EventHandler