+++ server/src/main/java/org/allaymc/server/network/processor/ingame/BlockEntityDataPacketProcessor.java
@@ -16,9 +16,10 @@ public class BlockEntityDataPacketProcessor extends PacketProcessor<BlockEntityD
     public void handleSync(EntityPlayer player, BlockEntityDataPacket packet, long receiveTime) {
         var pos = packet.getBlockPosition();
         var blockEntity = player.getDimension().getBlockEntity(pos.getX(), pos.getY(), pos.getZ());
-        ((BlockEntityBaseComponentImpl) ((BlockEntityImpl) blockEntity).getBaseComponent()).applyClientChange(player, packet.getData());
+        var baseComponent = (BlockEntityBaseComponentImpl) ((BlockEntityImpl) blockEntity).getBaseComponent();
+        baseComponent.applyClientChange(player, packet.getData());
         // Send new data to other viewers
-        blockEntity.sendBlockEntityToViewers();
+        baseComponent.sendBlockEntityToViewers();
     }
 
     @Override